pyproject.toml
[tool.poetry]
name = "digitalmarking"
version = "0.1.0"
description = ""
authors = ["Rinat <oblom.ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.2.1"
reportlab = "^4.4.0"
qrcode = "^8.2"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
